requirements.txt
aiofiles==23.2.1
annotated-types==0.6.0
anyio==3.7.1
backoff==2.2.1
black==23.10.1
boto3==1.28.78
botocore==1.31.78
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
fastapi==0.104.1
geographiclib==2.0
geopy==2.4.0
greenlet==3.0.1
h11==0.14.0
httptools==0.6.1
idna==3.4
jmespath==1.0.1
mypy-extensions==1.0.0
opencage==2.3.0
packaging==23.2
pathspec==0.11.2
piexif==1.1.3
Pillow==10.1.0
platformdirs==3.11.0
psycopg2==2.9.9
pydantic==2.4.2
pydantic_core==2.10.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
requests==2.31.0
s3transfer==0.7.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.27.0
typing_extensions==4.8.0
urllib3==2.0.7
uvicorn==0.24.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
